google_adk_agent/requirements.txt
google-adk>=0.3.0
requests>=2.28.0

